Refactor/general cleanup

+++ Scripts/compileMain.sh
@@ -1,7 +1,11 @@
-
 cd ../Main/Peripherals
-rm -rf Release
+if [ -d Release ]; then
+    rm -rf Release
+fi
 mkdir Release
-cd Release/
+cd Release
 cmake ..
 make Bootstrap
+
+cd ../../WebApplication
+npm install
